Quick Sort Structs
This was for a project for school.  This is a program that takes a file opens it reads and loads information into an array of structures.   It then sorts in ascending order based off salary in this use.  The functions take void*'s  why?  Because that's how I was told to do it.  Although it does make adapting the program to compare based off different values much easier and minimizes the tweaking needed.